+++ docs/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.1] - 2026-08-08
+### Changed
+- azure waf diagnostics flow through the session logger via azwaf v0.3.0's isolated slog
+  logger, replacing the logrus level workaround
+
 ## [0.8.0] - 2026-08-08
 ### Added
 - AhrefsBot, Applebot, DuckDuckBot and PerplexityBot web crawler providers, fetched via